Create new app users
Create register form function in views.py

Create register.html in template/users

Add the flash warining code to base.py when successfully sign up

Edit urls.py in travel app routes

users app handles the users who are going to sign up in the project for
diffrent apps. flash warning wont work because it's commented. and the
urls in travel is modified to route to '' instead of 'travel' as the
root